Calculate launch time instead of sending file name
Truncate all dependencies of function that sends file name to the
server and replace it with a new one which returns the moment of
starting parsing process of the current post.

There is no need in file names anymore since we start working with
databases. Moreover parsing launch time shows when exactly you start
to work with such post.